Add files via upload
sign_language.py contains the code to load the data and make ready for feeding to neural network
ASL_Recognition_challenge.ipynb is the python notebook with code and model building and evaluation. The sign_language.py file has been imported into this notebook.

The solution observed on kaggle is a bit different for loading the data.
The dataset is huge and hence can be downloaded from the competition page below. 
https://www.kaggle.com/fegadeharish/asl-recognition-using-deep-learning/data